task1_pipeline/requirements.txt
# Core ML libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
lightgbm>=3.3.0

# MLflow for experiment tracking
mlflow>=2.0.0

# FastAPI for model serving
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0

# Data processing
pyyaml>=6.0
joblib>=1.2.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0


# Optional: Docker support
# docker>=6.0.0
